Use customizable binary locations with diagnostics
This commit makes the binary locations for `coffee` and `sass`
customizable. It also provides the option of using Bundler
(http://bundler.io/) to run Sass.

In addition, the asset precompiler now performs a few checks before
compilation to make sure that the tools it needs exist. This also
gives better diagnostic messages.

The Sass option `--compass` is removed in this commit since not
everyone uses it.